previous fixes introduced a regression; According to the docs, you can not set up C::P::Cache without specifying a cache backend, but the tests did do this, and then registered a backend later.
This only worked because a croak in the C::P::Cache code was caught by an eval
and ignored, but the registering of the backend was done *before* the first
request to the cache.
Had a request gone to the cache first, it would have called the following code:

sub temporary_cache_backend {
my $c = shift;
die "FIXME - make up an in memory cache backend, that hopefully works well for the current engine";
}

This commit fixes the above issue by specifying a cache backend explicitly.
This is completely in line with what the documentation tells us to do.